+++ client/src/pages/create-post.js
@@ -1,0 +1,12 @@
+import CreatePost from "@/components/CreatePost";
+import React from "react";
+
+const PostCreate = () => {
+  return (
+    <>
+      <CreatePost />
+    </>
+  );
+};
+
+export default PostCreate;
