+++ client/src/styles/globals.css
@@ -1,0 +1,4 @@
+body {
+  font-family: "Rubik", sans-serif;
+  /* Other global styles here */
+}
